Improve performance

+++ DigitalSynthVRA8N/amp.h
@@ -10,6 +10,6 @@ public:
   }
 
   INLINE static int16_t clock(int16_t audio_input, uint8_t gain_control) {
-    return high_sbyte(audio_input) * (gain_control << 1);
+    return high_sbyte(audio_input) * static_cast<uint8_t>(gain_control << 1);
   }
 };
